fix: move extended WkWebView into plugin
